Laid framework for sbt and auto generating code and Intellij
Here is a rough start to the project. I've made it compatible with Intellij for easy importing (just make sure you have the Scala and SBT plugins).

The current main class is just for illustrative purposes. We are able to generate sources and then add them to the classpath at runtime.

`sbt compile` to compile the code, and generate sources
`sbt run` to run

Did a lot of researching for the past few days, but here's how it works:
In the `project/` directory we have a subproject that runs and generates all code.
`project/gen/scala/TokenKindGenerator` uses the `treehugger` library to output scala code.
In `project/Build.scala` we create this file as part of the build step. This is added to the top-level projects classpath.